fastapi-project/requirements.txt
fastapi
langchain
uvicorn
python-dotenv
openai